Adopt HiberniaModule bootstrap and 1.2.0 quick wins in treasury (PAR-261)
Migrate treasury's hand-rolled DI bootstrap to HiberniaModule.forPlugin(...),
activating the 1.2.0 defaults reconciliation. HiberniaModule now binds the
plugin, ConfigurationLoader, every @ConfigurationComponent, Message and the
command/resolver/listener multibinders; the kept TreasuryModule keeps only the
domain-service bindings, the concrete Treasury binding, the manually-parsed
config POJOs, the conditional LuckPerms binding and the two listener singletons.

- Treasury.onEnable(): build HiberniaModule with .handlers/.resolvers/.listeners,
  fetch LoggingConfiguration/DatabaseConfiguration via module.configuration(...)
  before the injector exists, and register listeners via ListenerManager.registerAll().
- Delete CommanderModule and EventsModule (their multibinders move into the builder).
- EconomyConfiguration.startingBalance double->BigDecimal; drop the redundant
  BigDecimal.valueOf wrappers in LedgerServiceImpl and EcoCommand.
- Replace the 14 hand-written help routes (TreasuryCommand, GovCommand) with a
  single paginated /treasury help [page] and /gov help [page] rendered by
  HelpGenerator over CommandManager.routeIndex(); CommandManager is injected as a
  Provider to break the handler<->CommandManager construction cycle. Remove the
  now-orphaned treasury.help and treasury.help.gov[.*] index message keys.
- ConfigReloadService: document that the in-place ConfigurationProcessor reload is
  intentional and must not be replaced with ConfigurationLoader.reload() (which
  swaps fresh instances and leaves the injected config singletons stale).
- TestConfigs: set the BigDecimal startingBalance field.

+++ treasury/src/main/java/io/paradaux/treasury/commands/EcoCommand.java
@@ -184,7 +184,7 @@ public class EcoCommand implements CommandHandler {
         UUID adminUuid = sender instanceof Player p ? p.getUniqueId() : TreasuryConstants.VIRTUAL_TREASURY_INITIATOR;
         ledgerService.adminReset(target.getUniqueId(), adminUuid);
 
-        BigDecimal startingBalance = Money.normalize(BigDecimal.valueOf(economyConfig.getStartingBalance()));
+        BigDecimal startingBalance = Money.normalize(economyConfig.getStartingBalance());
         String formattedAmount = accountService.formatAmount(startingBalance);
         message.send(sender, "treasury.eco.reset.success",
                 "target", target.getName(), "amount", formattedAmount);
